Factor out Series sorting tests to own file.
